pyproject.toml
[project]
requires-python = ">=3.13"
name = "aides-agri"
version = "0.1.0"
dependencies = [
  "beautifulsoup4 ~=4.13",
  "django ~=5.2",
  "django-admin-extra-buttons ~=2.1",
  "django-anymail[brevo] ~=13.0",
  "django-csp ~=4.0",
  "django-dsfr ~=2.1",
  "django-pgtrigger ~=4.15",
  "django-reversion ~=5.1",
  "django-tasks ~=0.6",
  "django-htmx ~=1.23",
  "django-two-factor-auth[phonenumberslite] ~=1.17",
  "gunicorn ~=23.0",
  "markdown ~=3.8",
  "markdownify ~=1.1",
  "mjml-python ~=1.3",
  "psycopg[binary] ~=3.0",
  "qrcode ~=7.4",
  "requests ~=2.32",
  "sentry-sdk[django] ~=2.26",
  "whitenoise ~=6.8",
]

[dependency-groups]
dev = [
  "djade",
  "django-browser-reload",
  "django-debug-toolbar",
  "ruff",
  "pytest",
  "pytest-cov",
  "pytest-django",
  "pytest-factoryboy",
  "requests-mock",
]

extraction = [
  # Core parsing & chunking
  "docling>=2.52.0",
  "docling-hierarchical-pdf>=0.1.2",
  "pdfminer>=20191125",
  "pypdf>=3.0.1",          # renamed from pypdf2 to match runtime import
  "unstructured[pdf]>=0.18.15",
  "trafilatura>=2.0.0",

  # LLM / structured extraction & tracking
  "ollama>=0.3.0",

  # Prompt loaders / community integrations (provides document loaders we use)
  "langchain-community>=0.3.29",

  # App & UX
  "streamlit>=1.49.1",
  "python-dotenv>=1.1.1",
  "tqdm>=4.67.1",
  "pandas>=2.3.3",

  # Notebooks / interactive benchmarks
  "ipykernel>=6.30.1",
  "ipywidgets>=8.1.7",

  # Data modelling
  "pydantic>=2.0.0",
]

veille = [
  "beautifulsoup4>=4.14.2",
  "browser-use>=0.8.0",
  "ipykernel>=6.30.1",
  "langchain-community>=0.3.31",
  "langchain-openai>=0.3.35",
  "langfuse>=3.8.0",
  "langgraph>=0.6.10",
  "langsmith>=0.4.34",
  "openai>=1.109.1",
  "selenium>=4.37.0",
]

[tool.coverage.run]
omit = ["**/tests/*"]

[tool.coverage.report]
show_missing = true
